implement webpack, use imports instead of namespaces, sort out webpack.config.js, amend package.json, npm install all the webpack things needed
