+++ package.json
@@ -5,13 +5,19 @@
   "main": "index.js",
   "dependencies": {
     "lite-server": "^2.5.4",
-    "package-lock.json": "^1.0.0",
-    "typescript": "^3.8.3"
+    "package-lock.json": "^1.0.0"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "ts-loader": "^7.0.1",
+    "typescript": "^3.8.3",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3"
+  },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "lite-server"
+    "start": "lite-server",
+    "build": "webpack"
   },
   "repository": {
     "type": "git",
