+++ src/component/postit-input.ts
@@ -1,55 +1,52 @@
-/// <reference path="base-component.ts" />
-/// <reference path="../decorators/autobind.ts" />
-/// <reference path="../util/validation.ts" />
-/// <reference path="../state/postit-column.ts" />
+import { ComponentPostit } from "./component-postit";
+import { postits } from "../state/postit-column";
+import { Postit } from "../models/postit";
 
-namespace App {
-  export class InputComponent extends Component<
-    HTMLDivElement,
-    HTMLFormElement
-  > {
-    titleInputElement: HTMLInputElement;
-    descriptionInputElement: HTMLInputElement;
-    valueInputElement: HTMLInputElement;
+export class InputComponent extends ComponentPostit<
+  HTMLDivElement,
+  HTMLFormElement
+> {
+  titleInputElement: HTMLInputElement;
+  descriptionInputElement: HTMLInputElement;
+  valueInputElement: HTMLInputElement;
 
-    constructor() {
-      super("project-input", "app", false, "user-input");
+  constructor() {
+    super("project-input", "app", false, "user-input");
 
-      this.titleInputElement = this.element.querySelector(
-        "#title"
-      ) as HTMLInputElement;
-      this.descriptionInputElement = this.element.querySelector(
-        "#description"
-      ) as HTMLInputElement;
-      this.valueInputElement = this.element.querySelector(
-        "#people"
-      ) as HTMLInputElement;
+    this.titleInputElement = this.element.querySelector(
+      "#title"
+    ) as HTMLInputElement;
+    this.descriptionInputElement = this.element.querySelector(
+      "#description"
+    ) as HTMLInputElement;
+    this.valueInputElement = this.element.querySelector(
+      "#people"
+    ) as HTMLInputElement;
 
-      this.configure();
-    }
+    this.configure();
+  }
 
-    configure() {
-      this.element.addEventListener("submit", this.submitHandler.bind(this));
-    }
+  configure() {
+    this.element.addEventListener("submit", this.submitHandler.bind(this));
+  }
 
-    private submitHandler(event: Event) {
-      event.preventDefault();
+  private submitHandler(event: Event) {
+    event.preventDefault();
 
-      const postit = new Postit(
-        this.titleInputElement.value,
-        this.descriptionInputElement.value,
-        +this.valueInputElement.value
-      );
+    const postit = new Postit(
+      this.titleInputElement.value,
+      this.descriptionInputElement.value,
+      +this.valueInputElement.value
+    );
 
-      const postitValidation = postits.addPostit(postit);
+    const postitValidation = postits.addPostit(postit);
 
-      if (postitValidation) {
-        this.titleInputElement.value = "";
-        this.descriptionInputElement.value = "";
-        this.valueInputElement.value = "";
-      } else {
-        alert(`invlaid input: ${"Invalid input"}`);
-      }
+    if (postitValidation) {
+      this.titleInputElement.value = "";
+      this.descriptionInputElement.value = "";
+      this.valueInputElement.value = "";
+    } else {
+      alert(`invlaid input: ${"Invalid input"}`);
     }
   }
 }
