+++ src/models/drag-drop.ts
@@ -1,12 +1,9 @@
-// Drag & Drop Interfaces
-namespace App {
-  export interface Draggable {
-    dragStartHandler(event: DragEvent): void;
-    dragEndHandler(event: DragEvent): void;
-  }
-  export interface DragTarget {
-    dragOverHandler(eventL: DragEvent): void;
-    dropHandler(eventL: DragEvent): void;
-    dragLeaveHandler(eventL: DragEvent): void;
-  }
+export interface Draggable {
+  dragStartHandler(event: DragEvent): void;
+  dragEndHandler(event: DragEvent): void;
+}
+export interface DragTarget {
+  dragOverHandler(eventL: DragEvent): void;
+  dropHandler(eventL: DragEvent): void;
+  dragLeaveHandler(eventL: DragEvent): void;
 }
