+++ src/models/postit.ts
@@ -1,17 +1,15 @@
-/// <reference path = "status.ts"/>
+import { Status } from "status";
 
-namespace App {
-  export class Postit {
-    id: string;
-    title: string;
-    description: string;
-    status: Status;
+export class Postit {
+  id: string;
+  title: string;
+  description: string;
+  status: Status;
 
-    constructor(t: string, d: string, status: Status) {
-      this.title = t.trim();
-      this.description = d.trim();
-      this.status = status;
-      this.id = Math.random().toString();
-    }
+  constructor(t: string, d: string, status: Status) {
+    this.title = t.trim();
+    this.description = d.trim();
+    this.status = status;
+    this.id = Math.random().toString();
   }
 }
