+++ src/models/status.ts
@@ -1,8 +1,6 @@
-namespace App {
-  export enum Status {
-    Unassigned,
-    Start,
-    Continue,
-    Stop,
-  }
+export enum Status {
+  Unassigned,
+  Start,
+  Continue,
+  Stop,
 }
