+++ src/util/validation.ts
@@ -1,44 +1,41 @@
-namespace App {
-  export interface Validatable {
-    value: string | number;
-    required?: boolean;
-    minLength?: number;
-    maxLength?: number;
+export interface Validatable {
+  value: string | number;
+  required?: boolean;
+  minLength?: number;
+  maxLength?: number;
 
-    minValue?: number;
-    maxValue?: number;
+  minValue?: number;
+  maxValue?: number;
+}
+
+export function validate(validatableInput: Validatable): boolean {
+  let isValid = true;
+
+  if (validatableInput.required) {
+    isValid = isValid && validatableInput.value.toString().trim().length !== 0;
   }
 
-  export function validate(validatableInput: Validatable): boolean {
-    let isValid = true;
-
-    if (validatableInput.required) {
-      isValid =
-        isValid && validatableInput.value.toString().trim().length !== 0;
-    }
-
-    if (validatableInput.minLength != null) {
-      isValid =
-        isValid &&
-        validatableInput.value.toString().trim().length >=
-          validatableInput.minLength;
-    }
-
-    if (validatableInput.maxLength != null) {
-      isValid =
-        isValid &&
-        validatableInput.value.toString().trim().length <=
-          validatableInput.maxLength;
-    }
-
-    if (validatableInput.minValue != null) {
-      isValid = isValid && validatableInput.value >= validatableInput.minValue;
-    }
-
-    if (validatableInput.maxValue != null) {
-      isValid = isValid && validatableInput.value <= validatableInput.maxValue;
-    }
-
-    return isValid;
+  if (validatableInput.minLength != null) {
+    isValid =
+      isValid &&
+      validatableInput.value.toString().trim().length >=
+        validatableInput.minLength;
   }
+
+  if (validatableInput.maxLength != null) {
+    isValid =
+      isValid &&
+      validatableInput.value.toString().trim().length <=
+        validatableInput.maxLength;
+  }
+
+  if (validatableInput.minValue != null) {
+    isValid = isValid && validatableInput.value >= validatableInput.minValue;
+  }
+
+  if (validatableInput.maxValue != null) {
+    isValid = isValid && validatableInput.value <= validatableInput.maxValue;
+  }
+
+  return isValid;
 }
